dbt_airflow/requirements-snowpark.txt

snowflake_snowpark_python[pandas]==1.12.0
dbt-core==1.7.7
dbt-snowflake==1.7.1
virtualenv==20.0.35
/tmp/astro_provider_snowflake-0.0.0-py3-none-any.whl
psycopg2-binary==2.9.6

snowflake-connector-python==3.7.0
snowflake-ml-python==1.2.1
astronomer-providers[all]==1.18.4
apache-airflow-providers-snowflake==5.3.0
